java.lang.NumberFormatException: Expected an int but was 5.5 at line 1 column 23408 path $.offset issue solved

+++ app/src/main/java/com/softaai/cityweather/data/remote/model/weather/WeatherResponse.kt
@@ -21,7 +21,7 @@ data class WeatherResponse(
     val daily: Daily,
     val alerts: List<Alert>,
     val flags: Flags,
-    val offset: Int
+    val offset: Double
 )
